requirements.txt
requests>=2.25.0
pandas>=1.3.0
openpyxl>=3.0.0
python-dotenv>=0.19.0
